% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare Treatment Estimate Methods}
\usage{
compare(...)
}
\arguments{
\item{...}{Objects of class \code{bonsaiforest}. We write them separated by a comma.}
}
\value{
List with a \code{data.frame} of the estimated subgroup treatment effects,
with the overall treatment effect and with the response type of our data.
}
\description{
Function to create a data set with the results for all the methods in order
to compare them.
}
\examples{
compare(naivepop_fit_surv, naive_fit_surv, elastic_net_fit_surv)
}
