% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareplot.R
\name{compareplot}
\alias{compareplot}
\title{Compare Forest Plots}
\usage{
compareplot(...)
}
\arguments{
\item{...}{Summary of \code{shrinkforest} objects separated by commas. There
should be at least two objects of this kind in the call of the function.}
}
\value{
Forest plot with all the fitted models.
}
\description{
Function to obtain a forest plot with all the different fitted models in
order to compare their performance.
}
\examples{
compareplot(
  summary(naive_fit_surv), summary(naivepop_fit_surv),
  summary(elastic_net_fit_surv)
)
}
