% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{cut_norm_quant}
\alias{cut_norm_quant}
\title{Helper for Cutting into Normal Quantiles}
\usage{
cut_norm_quant(x, prob, labels = letters[seq_along(c(prob, 1))])
}
\arguments{
\item{x}{(\code{numeric})\cr continuous covariate values.}

\item{prob}{(\code{numeric})\cr probabilities for the standard normal quantiles.}

\item{labels}{(\code{character})\cr corresponding labels for the resulting factor.}
}
\value{
The factor variable.
}
\description{
Helper for Cutting into Normal Quantiles
}
\keyword{internal}
