% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_dummy_bin1.R
\docType{data}
\name{design_dummy_bin1}
\alias{design_dummy_bin1}
\title{Design Dummy Binary subgroup x_1a}
\format{
A \code{matrix} with 0 and 1 indicating the subgroups of each of the observations.
}
\source{
This is an artificial example.
}
\usage{
design_dummy_bin1
}
\description{
The dummy design matrix from \code{example_data} considering just observations
from subgroup x_1a.
}
\keyword{datasets}
