% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix1.R
\docType{data}
\name{design_matrix1}
\alias{design_matrix1}
\title{Design Matrix Subgroup x_1a}
\format{
A \code{matrix} with 0 and 1 indicating the subgroups of each of the observations
and their interactions with the treatment variable.
}
\source{
This is an artificial example.
}
\usage{
design_matrix1
}
\description{
The design matrix from \code{example_data} considering just observations
from subgroup x_1a.
}
\keyword{datasets}
