% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_net_fit_bin.R
\docType{data}
\name{elastic_net_fit_bin}
\alias{elastic_net_fit_bin}
\title{Elastic Net Fit Binary}
\format{
A \code{elastic_net} object fitted with \code{example_data}, considering as subgroup
variables all the covariates x_1, ..., x_10 and \code{binary} response type.
}
\source{
This is an artificial example.
}
\usage{
elastic_net_fit_bin
}
\description{
A fitted model with the elastic_net function and binary data.
}
\keyword{datasets}
