% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_net_fit_surv.R
\docType{data}
\name{elastic_net_fit_surv}
\alias{elastic_net_fit_surv}
\title{Elastic Net Fit Survival}
\format{
A \code{elastic_net} object fitted with \code{example_data}, considering as subgroup
variables all the covariates x_1, ..., x_10 and \code{survival} response type.
}
\source{
This is an artificial example.
}
\usage{
elastic_net_fit_surv
}
\description{
A fitted model with the elastic_net function and survival data.
}
\keyword{datasets}
