% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_coef_bin1.R
\docType{data}
\name{est_coef_bin1}
\alias{est_coef_bin1}
\title{Estimated coefficients elastic net Binary}
\format{
A \code{matrix} with the coefficients of the elastic net model.
}
\source{
This is an artificial example.
}
\usage{
est_coef_bin1
}
\description{
The coefficients of the model \code{elastic_net_fit_bin} when lambda takes the
value that leads to minimal cross validation error.
}
\keyword{datasets}
