% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{horseshoe}
\alias{horseshoe}
\title{Bayesian Shrinkage Model Estimation}
\usage{
horseshoe(
  resp,
  trt,
  subgr,
  covars,
  data,
  resptype = c("survival", "binary"),
  status = NULL,
  chains = 4,
  iter = 2000,
  warmup = iter/2
)
}
\arguments{
\item{resp}{(\code{string})\cr the response variable name.}

\item{trt}{(\code{string})\cr the treatment variable name.The treatment variable
must be a factor with 2 levels where the first level is the control and the
second one the treatment.}

\item{subgr}{(\code{character})\cr vector with the name of the subgroup variables
from which we want to obtain the subgroup treatment effect.}

\item{covars}{(\code{character})\cr vector with the name of the variables that
we want to include in the model. The \code{subgr} variables have to be included
here.}

\item{data}{(\verb{data frame})\cr the data frame with the variables.}

\item{resptype}{(\code{string})\cr the type of data used. Can be "survival"
or "binary".}

\item{status}{(\code{string})\cr only for "survival" \code{resptype},
the status variable name in survival data.}

\item{chains}{(\code{scalar})\cr number of Markov chains in the brms model.
Default is 4.}

\item{iter}{(\code{scalar})\cr number of total iterations per chain in the brms
model (including warmup; defaults to 2000);}

\item{warmup}{(\code{scalar})\cr number of warmup (or burn-in) iterations in the
brms model. It should not be larger than \code{iter} and default is \code{iter/2}.}
}
\value{
List with \code{fit}, \code{model}, \code{resptype}, \code{data}, \code{alpha},
\code{design_matrix}, \code{design_dummy}, \code{y}, \code{subgr_names}.
}
\description{
Function to fit a bayesian shrinkage model with a regularized horseshoe prior
on the interaction coefficients between the subgrouping covariates and
the treatment.
}
\examples{
horseshoe("ev_pfs", "arm", c("x_1", "x_2"), c("x_1", "x_2", "x_3"),
example_data, "binary", chains = 1)
}
