% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_fit_bin.R
\docType{data}
\name{horseshoe_fit_bin}
\alias{horseshoe_fit_bin}
\title{Horseshoe Fit Binary}
\format{
A \code{horseshoe} object fitted with \code{example_data}, considering as subgroup
variables all the covariates x_1, ..., x_10 and \code{binary} response type.
}
\source{
This is an artificial example.
}
\usage{
horseshoe_fit_bin
}
\description{
A fitted model with the horseshoe function and binary data.
}
\keyword{datasets}
