% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_fit_surv.R
\docType{data}
\name{horseshoe_fit_surv}
\alias{horseshoe_fit_surv}
\title{Horseshoe Fit Survival}
\format{
A \code{horseshoe} object fitted with \code{example_data}, considering as subgroup
variables just the covariates x_1 and x_2 and \code{survival} response type.
}
\source{
This is an artificial example.
}
\usage{
horseshoe_fit_surv
}
\description{
A fitted model with the horseshoe function and survival data.
}
\keyword{datasets}
