% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahr_estimation.R
\name{km_fun}
\alias{km_fun}
\title{Helper Function to get Kaplan-Meier Estimate}
\usage{
km_fun(resp, status, data, times)
}
\arguments{
\item{resp}{(\code{string})\cr the response variable name.}

\item{status}{(\code{string})\cr only for "survival" \code{resptype},
the status variable name in survival data.}

\item{data}{(\code{data.frame})\cr data to use.}

\item{times}{(\code{numeric})\cr time points to use.}
}
\value{
Numeric vector with Kaplan-Meier estimates at \code{times}.
}
\description{
Helper Function to get Kaplan-Meier Estimate
}
\keyword{internal}
