% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lOR_estimation.R
\name{lOR_estimation}
\alias{lOR_estimation}
\title{Estimation of Log-Odds Ratio}
\usage{
lOR_estimation(X_subg, dummy_subg, est_coef)
}
\arguments{
\item{X_subg}{(\code{matrix})\cr the matrix with the subgroup covariates.}

\item{dummy_subg}{(\code{matrix})\cr the dummy matrix with the subgroup
covariates.}

\item{est_coef}{(\code{matrix})\cr the estimated coefficients from the fitted
model.}
}
\value{
Log-odds ratio of the studied subgroup.
}
\description{
Function to estimate the subgroup log-odds ratio for binary data.
}
\examples{
lOR_estimation(elastic_net_fit_bin$design_matrix,
 elastic_net_fit_bin$design_dummy, as.matrix(coef(elastic_net_fit_bin$fit,
 s = elastic_net_fit_bin$fit$lambda.min)))
}
