% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_fit_bin.R
\docType{data}
\name{naive_fit_bin}
\alias{naive_fit_bin}
\title{Naive Fit Binary}
\format{
A \code{naive} object fitted with \code{example_data}, considering as subgroup
variables all the covariates x_1, ..., x_10 and \code{binary} response type.
}
\source{
This is an artificial example.
}
\usage{
naive_fit_bin
}
\description{
A fitted model with the naive function and binary data.
}
\keyword{datasets}
