% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_fit_surv.R
\docType{data}
\name{naive_fit_surv}
\alias{naive_fit_surv}
\title{Naive Fit Survival}
\format{
A \code{naive} object fitted with \code{example_data}, considering as subgroup
variables all the covariates x_1, ..., x_10 and \code{survival} response type.
}
\source{
This is an artificial example.
}
\usage{
naive_fit_surv
}
\description{
A fitted model with the naive function and survival data.
}
\keyword{datasets}
