% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivepop.R
\name{naivepop}
\alias{naivepop}
\title{Naivepop function}
\usage{
naivepop(resp, trt, data, resptype, status = NULL)
}
\arguments{
\item{resp}{(\code{character}) \cr for the response variable name.}

\item{trt}{(\code{character}) \cr for the treatment variable name.}

\item{data}{(\verb{data frame}) \cr for the data frame with the variables.}

\item{resptype}{(\code{character}) \cr for the type of data used. Can be "survival"
or "binary".}

\item{status}{(\code{character}) \cr for the status variable name in survival data.}
}
\value{
List with fit, model, resptype, data
}
\description{
Function to fit the naivepop model to the data. This model is going to
estimate the subgroup treatment effects with the overall population
treatment effect.
}
\examples{
naivepop("futime", "trt", survival::myeloid, "survival", "death")
}
