% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivepop_fit.R
\docType{data}
\name{naivepop_fit}
\alias{naivepop_fit}
\title{Naivepop fit}
\format{
A \code{naivepop} object fitted with \code{example_data}.
}
\source{
This is an artificial example.
}
\usage{
naivepop_fit
}
\description{
A fitted model with the naivepop function.
}
\keyword{datasets}
