% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivepop_fit_bin.R
\docType{data}
\name{naivepop_fit_bin}
\alias{naivepop_fit_bin}
\title{Naivepop Fit Binary}
\format{
A \code{naivepop} object fitted with \code{example_data}, considering as subgroup
variables all the covariates x_1, ..., x_10 and \code{binary} response type.
}
\source{
This is an artificial example.
}
\usage{
naivepop_fit_bin
}
\description{
A fitted model with the naivepop function and binary data.
}
\keyword{datasets}
