% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivepop_fit_surv.R
\docType{data}
\name{naivepop_fit_surv}
\alias{naivepop_fit_surv}
\title{Naivepop Fit Survival}
\format{
A \code{naivepop} object fitted with \code{example_data}, considering as subgroup
variables all the covariates x_1, ..., x_10 and \code{survival} response type.
}
\source{
This is an artificial example.
}
\usage{
naivepop_fit_surv
}
\description{
A fitted model with the naivepop function and survival data.
}
\keyword{datasets}
