% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.compare.data}
\alias{plot.compare.data}
\title{Compare Forest Plots}
\usage{
\method{plot}{compare.data}(x, ...)
}
\arguments{
\item{x}{(\code{compare.data})\cr object with the data of treatment effects estimated
with the different methods.}

\item{...}{Additional arguments to plot.}
}
\value{
Forest plot with all the methods that are compared.
}
\description{
Function to obtain a forest plot with all the different fitted models in
order to compare their performance.
}
\examples{
plot(compare(naivepop_fit_surv, naive_fit_surv, elastic_net_fit_surv))
}
