% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.summary.elastic_net}
\alias{plot.summary.elastic_net}
\title{Forest plot Summary Elastic Net}
\usage{
\method{plot}{summary.elastic_net}(x, ...)
}
\arguments{
\item{x}{(\code{summary.elastic_net})\cr summary of a \code{elastic_net} object.}

\item{...}{Additional arguments to plot.}
}
\value{
Forest plot
}
\description{
Plot the forest plot of the summary of a \code{elastic_net} object.
}
\examples{
plot(summary(elastic_net_fit_surv))
}
