% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.summary.horseshoe}
\alias{plot.summary.horseshoe}
\title{Forest plot Summary Horseshoe}
\usage{
\method{plot}{summary.horseshoe}(x, ...)
}
\arguments{
\item{x}{(\code{summary.horseshoe})\cr summary of a \code{horseshoe} object.}

\item{...}{Additional arguments to plot.}
}
\value{
Forest plot
}
\description{
Plot the forest plot of the summary of a \code{horseshoe} object.
}
\examples{
plot(summary(horseshoe_fit_bin))
}
