% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.summary.naive}
\alias{plot.summary.naive}
\title{Forest plot Summary Naive}
\usage{
\method{plot}{summary.naive}(x, ...)
}
\arguments{
\item{x}{(\code{summary.naive})\cr summary of a \code{naive} object.}

\item{...}{Additional arguments to plot.}
}
\value{
Forest plot
}
\description{
Plot the forest plot of the summary of a \code{naive} object.
}
\examples{
plot(summary(naive_fit_surv))
}
