% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.elastic_net}
\alias{print.summary.elastic_net}
\title{Print Function for Elastic Net Summary}
\usage{
\method{print}{summary.elastic_net}(x, ...)
}
\arguments{
\item{x}{(\code{summary.elastic_net})\cr Summary of a \code{elastic_net} object.}

\item{...}{Other arguments of print.}
}
\value{
The summary of the object.
}
\description{
Function that prints the summary of a elastic_net object
}
\examples{
print(summary(elastic_net_fit_bin))
}
