% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.horseshoe}
\alias{print.summary.horseshoe}
\title{Print Function for Horseshoe Summary}
\usage{
\method{print}{summary.horseshoe}(x, ...)
}
\arguments{
\item{x}{(\code{summary.horseshoe})\cr Summary of a \code{horseshoe} object.}

\item{...}{Other arguments of print.}
}
\value{
The summary of the object.
}
\description{
Function that prints the summary of a horseshoe object
}
\examples{
print(summary(horseshoe_fit_bin))
}
