% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.naive}
\alias{print.summary.naive}
\title{Print Function for Naive Summary}
\usage{
\method{print}{summary.naive}(x, ...)
}
\arguments{
\item{x}{(\code{summary.naive})\cr Summary of a \code{naive} object.}

\item{...}{Other arguments of print.}
}
\value{
The summary of the object.
}
\description{
Function that prints the summary of a naive object
}
\examples{
print(summary(naive_fit_bin))
}
