% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.naivepop}
\alias{print.summary.naivepop}
\title{Print Function for Naivepop Summary}
\usage{
\method{print}{summary.naivepop}(x, ...)
}
\arguments{
\item{x}{(\code{summary.naivepop})\cr Summary of a \code{naivepop} object.}

\item{...}{Other arguments of print.}
}
\value{
The summary of the object.
}
\description{
Function that prints the summary of a naivepop object
}
\examples{
print(summary(naivepop_fit_bin))
}
