% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkforest-package.R
\docType{package}
\name{shrinkforest-package}
\alias{shrinkforest}
\alias{shrinkforest-package}
\title{shrinkforest: Shrinkage based Forest Plots}
\description{
Subgroup analyses are routinely performed in clinical trial analyses. From a methodological perspective, two key issues of subgroup analyses are multiplicity (even if only predefined subgroups are investigated) and the low sample sizes of subgroups which lead to highly variable estimates. This package implements subgroup estimates based on Bayesian shrinkage priors, as well as penalized likelihood inference.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/shrinkforest/}
  \item Report bugs at \url{https://github.com/insightsengineering/shrinkforest/issues}
}

}
\author{
\strong{Maintainer}: Mar Vazquez Rabunal \email{mar.vazquez_rabunal@roche.com}

}
\keyword{internal}
