% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.naive}
\alias{summary.naive}
\title{Summary Naive}
\usage{
\method{summary}{naive}(object, conf = 0.95, ...)
}
\arguments{
\item{object}{(\code{naive})\cr the naive object.}

\item{conf}{(\code{scalar})\cr the confidence level of the intervals. Default is
0.95.}

\item{...}{Arguments of summary.}
}
\value{
\code{trt_effect}
}
\description{
Function to obtain the naive subgroup treatment effects of an object fitted
with the \code{naive} function.
}
\examples{
summary(naive_fit_surv)
}
