% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.naivepop}
\alias{summary.naivepop}
\title{Summary Function}
\usage{
\method{summary}{naivepop}(object, ...)
}
\arguments{
\item{object}{(\code{shrinkforest})\cr the shrinkforest object.}

\item{...}{\cr arguments of summary.}
}
\value{
The estimated subgroup treatment effects.
}
\description{
Function to obtain the subgroup treatment effect of a shrinkforest object.
}
\examples{
summary(naivepop_fit_surv)
}
