% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_curves.R
\name{survival_curves}
\alias{survival_curves}
\title{Average Survival Curves}
\usage{
survival_curves(x, h0, est_coef)
}
\arguments{
\item{x}{(\code{matrix})\cr the matrix with the subgroup covariates.}

\item{h0}{(\code{numeric})\cr the vector with the cumulative baseline hazard.}

\item{est_coef}{(\code{matrix})\cr the estimated coefficients from the fitted
model.}
}
\value{
The survival probabilities at the sorted event times.
}
\description{
Function to obtain the average survival curve from all individual
survival curves.
}
\examples{
survival_curves(
  elastic_net_surv$x, elastic_net_surv$h0,
  elastic_net_surv$est_coef
)
}
