% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Little Test Function doing Stuff}
\usage{
test(bla, bli, blu, ...)
}
\arguments{
\item{bla}{(\code{string})\cr for bla stuff.}

\item{bli}{(\code{numeric})\cr for bli stuff.}

\item{blu}{(\code{list})\cr for blu.}

\item{...}{additional arguments.}
}
\value{
List with bla, bli, blu.
}
\description{
This is testing documenting arguments. Because it is cool.
}
\examples{
test("bli", 2, list(1, 4, 5))
}
