% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_net_surv.R
\docType{data}
\name{elastic_net_surv}
\alias{elastic_net_surv}
\title{H0, Coefficients Elastic Net Survival Model and Matrices}
\format{
A \code{list} with three elements:
\itemize{
\item \code{h0}: the cumulative baseline hazard.
\item \code{est_coef}: the estimated coefficients of the fitted model.
\item \code{x}: the design matrix of the model.
\item \code{dummy1}: the dummy matrix of the model for subgroup x_1a.
\item \code{design1}: the design matrix of the model for subgroup x_1a.
}
}
\source{
This is an artificial example.
}
\usage{
elastic_net_surv
}
\description{
List with the cumulative baseline hazard, with the estimated coefficients
of the model fitted with an elastic net and survival data, with the general
design matrix and with the design and dummy matrices for subgroup x_1a.
}
\keyword{datasets}
