% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
A \code{data.frame} with 1000 rows and 14 columns:
\itemize{
\item \code{id}: Integer variable with the id of the patient.
\item \code{arm}: Factor variable with 2 levels: control and treatment.
\item \code{x_1}, \code{x_2}, \code{x_3}, \code{x_4}, \code{x_5}, \code{x_6}, \code{x_7}, \code{x_8}, \code{x_9}, \code{x_10}:
Factor variables that represent covariates of the patients
(e.g. sex, region, race, etc.).
\item \code{tt_pfs}: Continuous variable with the progression free survival time
of each patient.
\item \code{ev_pfs}: Binary variable with the progression free survival status
of each patient. 0 indicates that the event of interest did not happen
and 1 indicates that it did.
}
}
\source{
This is an artificial dataset.
}
\usage{
example_data
}
\description{
A simulated data set for a realistic scenario of a clinical trial.
}
\keyword{datasets}
