% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lOR_estimation.R
\name{lor_estimation}
\alias{lor_estimation}
\title{Estimation of Log-Odds Ratio}
\usage{
lor_estimation(x_subg, dummy_subg, est_coef)
}
\arguments{
\item{x_subg}{(\code{matrix})\cr the matrix with the subgroup covariates.}

\item{dummy_subg}{(\code{matrix})\cr the dummy matrix with the subgroup
covariates.}

\item{est_coef}{(\code{matrix})\cr the estimated coefficients from the fitted
model.}
}
\value{
Log-odds ratio of the studied subgroup.
}
\description{
Function to estimate the subgroup log-odds ratio for binary data.
}
\examples{
lor_estimation(design_matrix1, design_dummy1, est_coef_bin1)
}
