% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.naive}
\alias{summary.naive}
\title{Summary Naive}
\usage{
\method{summary}{naive}(object, conf = 0.95, ...)
}
\arguments{
\item{object}{(\code{naive})\cr the naive object.}

\item{conf}{(\code{scalar})\cr the confidence level of the intervals. Default is
0.95.}

\item{...}{Arguments of summary.}
}
\value{
Object of class \code{summary.elastic_net} which is a \code{list} with a
\code{data.frame} with 4 columns (the subgroup variables, the estimated
treatment effect and the low and high bounds of the confidence interval of the
treatment effect), the \code{resptype} and the confidence level.
}
\description{
Function to obtain the naive subgroup treatment effects of an object fitted
with the \code{naive} function.
}
\examples{
summary(naive_fit_surv)
}
