% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.naivepop}
\alias{summary.naivepop}
\title{Summary Naivepop Function}
\usage{
\method{summary}{naivepop}(object, ...)
}
\arguments{
\item{object}{(\code{naivepop})\cr the naivepop object.}

\item{...}{Arguments of summary.}
}
\value{
Object of class \code{summary.naivepop} which is a \code{list} with the
estimated subgroup treatment effects and the \code{resptype}.
}
\description{
Function to obtain the overall subgroup treatment effect of a
naivepop object.
}
\examples{
summary(naivepop_fit_surv)
}
